Meshflow/requirements.dev.txt
# Testing
pytest~=8.0
pytest-django~=4.8
pytest-cov~=4.1
pytest-mock~=3.12
pytest-factoryboy~=2.5
factory-boy~=3.3
coverage~=7.4

# Development tools
black~=24.2
flake8~=7.0
isort~=5.13

# Include production requirements
-r requirements.txt
